refactor: auth 관련 폴더 domain으로 이전
